updated many bugs ,added new pages

+++ src/admin_panel/partials/draft-trow.php
@@ -23,22 +23,13 @@
    echo ucwords($ca['name']);
   }
        ?></td>
-<td class="px-2 text-center font-semibold"> <a href="see-comments?com_id=<?php echo $id; ?>" title="comments" > <?php
-                        //Sql Query
-    $getcom = "SELECT * FROM tbl_comments WHERE post_id=$id";
-                        //Execute Query
-                        $comnum = mysqli_query($conn, $getcom);
-                        //Count Rows
-                        $countcom = mysqli_num_rows($comnum);
-                        echo $countcom;
-                    ?><i class="ml-2 fa-solid fa-message"></i></a></td>
 
 <td class="px-2 text-center break-all"><?php echo $date; ?></td>
 
 <td class="break-words"> 
-<a href="../blog_article?post_key=<?php echo $id; ?>" class="p-1 bg-green-500 text-white rounded font-semibold" target="_blank">VIEW</a>
+<a href="see-drafts?publish_key=<?php echo $id; ?>" class="p-1 bg-green-500 text-white rounded font-semibold" target="_blank">Publish</a>
 <a href="edit_post?id=<?php echo $id; ?>" class="p-1 bg-[#2271B1] text-white rounded ml-2 font-semibold">EDIT</a>
-<a href="delete?id=<?php echo $id; ?>&cover_image=<?php echo $c_img; ?>"  class="p-1 bg-[#A23445] text-white rounded ml-2 font-semibold">DELETE</a>
+<a href="delete?id=<?php echo $id; ?>&cover_image=<?php echo $c_img; ?>" onclick="confirm('Are you sure to Delete this Article?')" class="p-1 bg-[#A23445] text-white rounded ml-2 font-semibold">DELETE</a>
   
 </td>
 </tr>
